+++ noodles-cram/src/data_container/slice/builder.rs
@@ -114,9 +114,8 @@ impl Builder {
                     .expect("missing reference sequence")
                     .expect("invalid reference sequence");
 
-                let mut hasher = Md5::new();
-                hasher.update(&reference_sequence[start..=end]);
-                <[u8; 16]>::from(hasher.finalize())
+                let sequence = &reference_sequence[start..=end];
+                calculate_normalized_sequence_digest(sequence)
             }
             _ => [0; 16],
         };
@@ -262,5 +261,54 @@ fn update_substitution_features(substitution_matrix: &SubstitutionMatrix, featur
             }
             _ => {}
         }
+    }
+}
+
+// _Sequence Alignment/Map Format Specification_ (2021-06-03) § 1.3.2 "Reference MD5 calculation"
+fn calculate_normalized_sequence_digest(sequence: &[u8]) -> [u8; 16] {
+    let mut hasher = Md5::new();
+
+    for &b in sequence {
+        // "All characters outside of the inclusive range 33 ('!') to 126 ('~') are stripped out."
+        if b.is_ascii_graphic() {
+            // "All lowercase characters are converted to uppercase."
+            hasher.update(&[b.to_ascii_uppercase()]);
+        }
+    }
+
+    hasher.finalize().into()
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_calculate_normalized_sequence_digest() {
+        assert_eq!(
+            calculate_normalized_sequence_digest(b"ACGT"),
+            [
+                0xf1, 0xf8, 0xf4, 0xbf, 0x41, 0x3b, 0x16, 0xad, 0x13, 0x57, 0x22, 0xaa, 0x45, 0x91,
+                0x04, 0x3e
+            ]
+        );
+
+        assert_eq!(
+            calculate_normalized_sequence_digest(b"ACgt"),
+            [
+                0xf1, 0xf8, 0xf4, 0xbf, 0x41, 0x3b, 0x16, 0xad, 0x13, 0x57, 0x22, 0xaa, 0x45, 0x91,
+                0x04, 0x3e
+            ]
+        );
+
+        // _Sequence Alignment/Map Format Specification_ (2021-06-03) § 1.3.2 "Reference MD5
+        // calculation"
+        assert_eq!(
+            calculate_normalized_sequence_digest(b"ACGTACGTACGTACGTACGTACGT...12345!!!"),
+            [
+                0xdf, 0xab, 0xdb, 0xb3, 0x6e, 0x23, 0x9a, 0x6d, 0xa8, 0x89, 0x57, 0x84, 0x1f, 0x32,
+                0xb8, 0xe4
+            ]
+        );
     }
 }